Fix undo/redo bug on chord items
The bug was caused by the item init function for chord items which updated filters when an item was edited.

The filter update got reported first to the history manager before the start/end change could be reported.

Essentially, this meant the new state with the new start/end was stored as the original state of the item and thus when we undo, this state was restored instead of the actual old state with the old start/end.

I solved this by removing item init functions completely as they were only used for chord item filters and filters are not being used.